fitness_system/backend/requirements.txt
Django>=5.2,<6.1
 djangorestframework>=3.16,<3.17
 django-cors-headers>=4.9,<5.0
 pymysql>=1.1,<2.0
